refactor(server): split admin tokens and middleware
