+++ backend/internal/server/middleware_cors.go
@@ -1,0 +1,48 @@
+package server
+
+import (
+	"net/http"
+	"strings"
+)
+
+// corsOrigin returns the configured Access-Control-Allow-Origin, treating an
+// empty value as the "*" default (an empty .env line must not disable CORS).
+func (s *Server) corsOrigin() string {
+	origin := strings.TrimSpace(s.cfg.Load().CORSAllowedOrigin)
+	if origin == "" {
+		return "*"
+	}
+	return origin
+}
+
+// corsMiddleware answers CORS preflights on the /v1/* API surface and stamps
+// the allow headers on /v1/* responses. An OPTIONS request for any /v1/*
+// path is answered with 204 before the route table sees it (so unknown
+// /v1/* subpaths still get a clean preflight, matching the reference
+// proxy-freebuff OPTIONS → 204). Admin routes pass through untouched.
+func (s *Server) corsMiddleware(next http.Handler) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		if strings.HasPrefix(r.URL.Path, "/v1/") {
+			h := w.Header()
+			origin := s.corsOrigin()
+			h.Set("Access-Control-Allow-Origin", origin)
+			// When the origin is pinned (not "*"), vary on Origin so caches
+			// never serve the pinned header to a different requester.
+			if origin != "*" {
+				h.Add("Vary", "Origin")
+			}
+			// The allow-list must cover every credential/header the auth path
+			// accepts (x-api-key, anthropic-api-key) and the Anthropic beta
+			// header real clients carry — a browser preflight fails silently
+			// otherwise ("header not allowed") and the real request never
+			// fires.
+			h.Set("Access-Control-Allow-Headers", "Content-Type, Authorization, x-api-key, anthropic-api-key, anthropic-version, anthropic-beta")
+			h.Set("Access-Control-Allow-Methods", "POST, GET, OPTIONS")
+			if r.Method == http.MethodOptions {
+				w.WriteHeader(http.StatusNoContent)
+				return
+			}
+		}
+		next.ServeHTTP(w, r)
+	})
+}
